% Auto-generated: do not edit by hand
\name{dropdown}

\alias{dropdown}

\title{Dropdown component}

\description{

}

\usage{
dropdown(id=NULL, className=NULL, clearable=NULL, disabled=NULL,
multi=NULL, options=NULL, placeholder=NULL, searchable=NULL,
style=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{className}{Character. }

\item{clearable}{Logical. }

\item{disabled}{Logical. }

\item{multi}{Logical. }

\item{options}{List of lists containing elements 'value', 'label'.
those elements have the following types:
  - value (character | numeric; required)
  - label (character; optional)s. }

\item{placeholder}{Character. }

\item{searchable}{Logical. }

\item{style}{Named list. }

\item{value}{List of lists containing elements 'value', 'label'.
those elements have the following types:
  - value (character | numeric; required)
  - label (character; optional)s | lists containing elements 'value', 'label'.
those elements have the following types:
  - value (character | numeric; required)
  - label (character; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

